get rid of configsettings
this layer of indirection was mainly useful for testing, but we test a
bit differently with the new config code (testing the parent classes at
a fundamental level as opposed to every individual property
implementation).

we may need to bring this back when we get to the command-line args.
