+++ src/main/kotlin/org/jitsi/videobridge/config/CommandLineArgsConfigSource.kt
@@ -19,9 +19,8 @@ package org.jitsi.videobridge.config
 import org.jitsi.utils.config.ConfigSource
 import kotlin.reflect.KClass
 
-class CommandLineArgsConfigSource : ConfigSource {
+class CommandLineArgsConfigSource(private val commandLineArgs: Array<String>) : ConfigSource {
     override val name: String = "command line args"
-    private val commandLineArgs = ConfigSupplierSettingsk.commandLineArgsSupplier()
 
     override fun <T : Any> getterFor(valueType: KClass<T>): (String) -> T {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
